+++ modules/tinymce/src/core/test/ts/browser/keyboard/EnterKeyCeFalseTest.ts
@@ -74,4 +74,32 @@ describe('browser.tinymce.core.keyboard.EnterKeyCeFalseTest', () => {
     TinyAssertions.assertContent(editor, '<p><span contenteditable="false">x</span></p>');
     assert.equal(editor.selection.getNode().nodeName, 'SPAN');
   });
+
+  it('TINY-9101: Pressing Enter on a cE=false block should do nothing', () => {
+    const editor = hook.editor();
+    editor.setContent('<p>First</p><p contenteditable="false">Second</p><p>Third</p>');
+    TinySelections.select(editor, 'p:eq(1)', [ ]);
+    pressEnter(editor);
+    TinyAssertions.assertContent(editor, '<p>First</p><p contenteditable="false">Second</p><p>Third</p>');
+    assert.equal(editor.selection.getNode().nodeName, 'P');
+  });
+
+  it('TINY-9101: Pressing Enter on a cE=false pre should do nothing', () => {
+    const editor = hook.editor();
+    editor.setContent('<p>First</p><pre contenteditable="false">Second</pre><p>Third</p>');
+    TinySelections.select(editor, 'pre', [ ]);
+    pressEnter(editor);
+    TinyAssertions.assertContent(editor, '<p>First</p><pre contenteditable="false">Second</pre><p>Third</p>');
+    assert.equal(editor.selection.getNode().nodeName, 'PRE');
+  });
+
+  it('TINY-9101: Enter after selecting across paragraphs before a cE=false span should not delete cE=false span', () => {
+    const editor = hook.editor();
+    editor.setContent('<p>first</p><p>second<span contenteditable="false">2</span></p>');
+    // Select across the two paragraphs, but stop before the cef span
+    TinySelections.setSelection(editor, [ 0, 0 ], 'fir'.length, [ 1, 0 ], 'sec'.length);
+    pressEnter(editor);
+    TinyAssertions.assertContent(editor, '<p>fir</p><p>ond<span contenteditable="false">2</span></p>');
+    assert.equal(editor.selection.getNode().nodeName, 'P');
+  });
 });