+++ modules/mcagar/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Added `promotion` set to `false` as a default setting.
+
 ## 8.0.0 - 2022-03-03
 
 ### Added