+++ modules/mcagar/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-### Added
+### Improved
 - Added `promotion` set to `false` as a default setting.
 
 ## 8.0.0 - 2022-03-03