+++ modules/mcagar/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 8.1.0 - 2022-09-08
+
 ### Changed
 - The new `promotion` option added in TinyMCE 6.2 is set to `false` by default.